Safari-611.1.10/LayoutTests/svg/dynamic-updates/SVGTextElement-svgdom-dy-prop.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGTextElement-svgdom-dy-prop.js
// [Expected rendering result] text message shifted to 0x20 - and a series of PASS messages

description("Tests dynamic updates of the 'dy' attribute of the SVGTextElement object")
createSVGTestCase();

var textElement = createSVGElement("text");
textElement.setAttribute("x", "0");
textElement.setAttribute("y", "50");
textElement.setAttribute("dy", "0");
textElement.textContent="Text content";

rootSVGElement.appendChild(textElement);
shouldBe("textElement.dy.baseVal.getItem(0).value", "0");

function repaintTest() {
    textElement.dy.baseVal.getItem(0).value = -30;
    shouldBe("textElement.dy.baseVal.getItem(0).value", "-30");
    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
